+++ docs/index.md
@@ -25,7 +25,7 @@ api/data_collector
 :glob:
 :caption: Tutorials
 tutorials/**/index
-tutorials/q_learning_torch
+tutorials/behavioral_cloning
 ```
 
 ```{toctree}
